+++ app/(blog)/admin/layout.tsx
@@ -1,6 +1,6 @@
 import AuthContext from "@/app/components/context/AuthContext";
 import ToastContext from "@/app/components/context/ToastContext";
-import Header from "@/app/components/Header";
+import Header from "@/app/components/section/Header";
 import Footer from "@/app/components/Footer";
 import { Metadata } from "next";
