+++ app/components/section/Header.tsx
@@ -1,0 +1,42 @@
+import Image from "next/image";
+import Link from "next/link";
+
+type HeaderProps = {
+  isTopPage?: boolean;
+};
+
+const Header: React.FC<HeaderProps> = ({ isTopPage }) => {
+  const webSiteTitle = process.env.NEXT_PUBLIC_WEBSITE_TITLE;
+
+  return (
+    <header>
+      <div className="flex items-center justify-center">
+        <Link href="/">
+          {isTopPage ? (
+            <h1>
+              <Image
+                src="/logo_blog.png"
+                alt={webSiteTitle || "サイトタイトル"}
+                width={350}
+                height={90}
+                priority
+                className="w-[220px] h-auto md:w-[350px]"
+              />
+            </h1>
+          ) : (
+            <Image
+              src="/logo_blog.png"
+              alt={webSiteTitle || "サイトタイトル"}
+              width={350}
+              height={90}
+              priority
+              className="w-[220px] h-auto md:w-[350px]"
+            />
+          )}
+        </Link>
+      </div>
+    </header>
+  );
+};
+
+export default Header;
