+++ app/page.tsx
@@ -1,4 +1,4 @@
-import Header from "./components/Header";
+import Header from "./components/section/Header";
 import TopCategories from "./components/section/TopCategories";
 import TopNewArticles from "./components/section/TopNewArticles";
 import Footer from "./components/Footer";
